+++ src/main/kotlin/de/mecrytv/earthcore/database/annotations/Column.kt
@@ -2,4 +2,4 @@ package de.mecrytv.earthcore.database.annotations
 
 @Target(AnnotationTarget.PROPERTY, AnnotationTarget.FIELD)
 @Retention(AnnotationRetention.RUNTIME)
-annotation class Column(val name: String = "", val length: Int = 255)
+annotation class Column(val name: String = "", val length: Int = 255, val text: Boolean = false)