+++ src/main/kotlin/de/mecrytv/earthcore/logging/api/LogSink.kt
@@ -1,0 +1,10 @@
+package de.mecrytv.earthcore.logging.api
+
+interface LogSink {
+
+    val name: String
+
+    fun accept(entry: LogEntry)
+
+    fun close()
+}