+++ src/main/kotlin/de/mecrytv/earthcore/logging/api/LogbookProvider.kt
@@ -1,0 +1,10 @@
+package de.mecrytv.earthcore.logging.api
+
+import org.bukkit.plugin.Plugin
+
+interface LogbookProvider {
+
+    fun of(plugin: Plugin): Logbook
+
+    fun sinks(): List<String>
+}